moving converters, fixing spacing issues
