+++ src/modules/fancyzones/editor/FancyZonesEditor/EditorOverlay.xaml.cs
@@ -2,22 +2,10 @@
 // The Microsoft Corporation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using FancyZonesEditor.Models;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net.NetworkInformation;
-using System.Security.RightsManagement;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
 using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Shapes;
+using FancyZonesEditor.Models;
 
 namespace FancyZonesEditor
 {
@@ -70,6 +58,7 @@ namespace FancyZonesEditor
         }
 
         public static EditorOverlay Current;
+
         public EditorOverlay()
         {
             InitializeComponent();
